Added module for PageType classes. Use `app` as a folder for classes
