+++ manifest.json
@@ -56,6 +56,7 @@
         "assets/uuid/validate.js",
         "images/b2d_logo_16.png",
         "src/app/release.js",
+        "src/bandcamp/app/page-type.js",
         "src/bandcamp/modules/bandcamp.js",
         "src/bandcamp/modules/html.js",
         "src/bandcamp/modules/main.js",
@@ -64,7 +65,7 @@
         "src/bandcamp/pages/page-music.js",
         "src/bandcamp/styles.css",
         "src/config.js",
-        "src/discogs/classes/discogs-csv.js",
+        "src/discogs/app/discogs-csv.js",
         "src/discogs/modules/discogs.js",
         "src/discogs/modules/genres.js",
         "src/modules/chrome.js",
