+++ src/bandcamp/modules/main.js
@@ -1,6 +1,6 @@
 import { logStorage } from "../../modules/storage.js";
 import { injectCSSFile, injectJSFile } from "../../modules/utils.js";
-import { PageTypeDetector } from "./bandcamp.js";
+import { PageTypeDetector } from "../app/page-type.js";
 import { setupPageAlbum } from "../pages/page-album.js";
 import { setupPageMusic } from "../pages/page-music.js";
