+++ src/popup/tabs/download_tab.js
@@ -1,5 +1,5 @@
 import { Release } from "../../app/release.js";
-import { DiscogsCsv } from "../../discogs/classes/discogs-csv.js";
+import { DiscogsCsv } from "../../discogs/app/discogs-csv.js";
 import { downloadCsv, objectsToCsv } from "../../modules/csv.js";
 import { isEmptyArray } from "../../modules/utils.js";
 import { disable, enable } from "../../modules/html.js";
