+++ src/popup/tabs/storage_tab.js
@@ -1,4 +1,4 @@
-import { DiscogsCsv } from "../../discogs/classes/discogs-csv.js";
+import { DiscogsCsv } from "../../discogs/app/discogs-csv.js";
 import { downloadCsv, objectsToCsv } from "../../modules/csv.js";
 import { createElementFromHTML, hasDataAttribute, setDataAttribute } from "../../modules/html.js";
 import { clearStorage, clearStorageByKey, findAllReleases, findReleasesByUrls } from "../../modules/storage.js";
